requirements.txt
flask==3.0.2
python-dotenv==1.0.1
flask-sqlalchemy==3.1.1
flask-login==0.6.3
werkzeug==3.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9  # untuk PostgreSQL
